Move codecs, update in-module grib example
